Updated `VariableInitializationCheck` with documentation and support when working with new `Var` helper methods.
